// src/interfaces/phr/Encounter.type.ts
export interface CoverageType {
    system_identifier: string
    identifier: string
    subscriberId: string
    status: string
    system: string
    system_type: string
    code_type: string
    display_type: string
    system_relationship: string
    code_relationship: string
    display_relationship: string
    start_period: string
    end_period: string
    reference_payor: any
    system_class: string
    code_class: string
    value_class: string
    name_class: string
    value_class_class: string
    reimbursementAmount_class: string
    reference_contract: any
    identifier_contract: string
    display_contract: any
    hcode_Coverage: string
    hosname: string
    vn_Coverage: string
    HN: string
    system_clinicalStatus: string
    code_clinicalStatus: string
    display_clinicalStatus: string
    system_verificationStatus: string
    code_verificationStatus: string
    display_verificationStatus: string
    system_category: string
    code_category: string
    display_category: string
    system_severity: string
    code_severity: string
    display_severity: string
    system_code: string
    code_code: string
    display_code: string
    text_code: string
    system_bodySite: string
    code_bodySite: string
    display_bodySite: string
    text_bodySite: string
    recordedDate_bodySite: string
  }
  export interface VitalSignsType {
    status_body_weight: string
    value_body_weight: string
    unit_body_weight: string
    status_body_height: string
    value_body_height: string
    unit_body_height: string
    status_body_temp: string
    value_body_temp: string
    unit_body_temp: string
    status_bp_systolic: string
    value_bp_systolic: string
    unit_bp_systolic: string
    text_bp_systolic: string
    status_bp_diastolic: string
    value_bp_diastolic: string
    unit_bp_diastolic: string
    text_bp_diastolic: string
  }
  